[java]  Moving (firefox) Executable out of internal package to the main one, but making it package private
