+++ java/client/src/org/openqa/selenium/firefox/FirefoxBinary.java
@@ -29,7 +29,6 @@ import com.google.common.collect.Iterables;
 
 import org.openqa.selenium.Platform;
 import org.openqa.selenium.WebDriverException;
-import org.openqa.selenium.firefox.internal.Executable;
 import org.openqa.selenium.io.FileHandler;
 import org.openqa.selenium.os.CommandLine;
 import org.openqa.selenium.os.ExecutableFinder;
